Enable FindBugs in tests.
This would have caught the unused variable that was fixed by #1082.

Changes in this commit:

  - Enable FindBugs in tests in build.gradle.

  - Fix some existing minor FindBugs issues in tests.

  - Suppress some warnings in all test classes that are less important in test
    code, such as requiring nested classes to be static whenever possible.

  - Enable nullness-related FindBugs warnings in tests, since we don't run the
    Checker Framework on test code.

+++ impl_core/src/test/java/io/opencensus/implcore/stats/ViewManagerImplTest.java
@@ -97,7 +97,7 @@ public class ViewManagerImplTest {
   private static final Cumulative CUMULATIVE = Cumulative.create();
 
   private static final double EPSILON = 1e-7;
-  private static final int MILLIS_PER_SECOND = 1000;
+  private static final long MILLIS_PER_SECOND = 1000;
   private static final Duration TEN_SECONDS = Duration.create(10, 0);
   private static final Interval INTERVAL = Interval.create(TEN_SECONDS);
